Testing
Testing revealed you cant automatically cast a long to an integer so i fixed it

+++ src/armor/Factory_Heavy_Armor.java
@@ -17,15 +17,15 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 			// Reads JSON file for all attributes of the heavy head armor
 			// Overengineered, yes, but more memory-efficient
 			return new Head_Armor("heavy", 
-					(Integer)head_armor_json.get("BLUNT_RESISTANCE"),
-					(Integer)head_armor_json.get("CUT_RESISTANCE"), 
-					(Integer)head_armor_json.get("COVERAGE"), 
+					((Long)head_armor_json.get("BLUNT_RESISTANCE")).intValue(),
+					((Long)head_armor_json.get("CUT_RESISTANCE")).intValue(), 
+					((Long)head_armor_json.get("COVERAGE")).intValue(), 
 					// Attributes
 					new HashMap<String, Integer> () {{
 						JSONObject attributes_json = (JSONObject)head_armor_json.get("ATTRIBUTES");
 						for (Iterator it = attributes_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)attributes_json.get(key));
+							put(key, ((Long)attributes_json.get(key)).intValue());
 						}
 					}},
 					// Combat skills
@@ -33,7 +33,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject combat_skills_json = (JSONObject)head_armor_json.get("COMBAT_SKILLS");
 						for (Iterator it = combat_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)combat_skills_json.get(key));
+							put(key, ((Long)combat_skills_json.get(key)).intValue());
 						}
 					}},
 					// Weapon skills
@@ -41,7 +41,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject weapon_skills_json = (JSONObject)head_armor_json.get("WEAPON_SKILLS");
 						for (Iterator it = weapon_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)weapon_skills_json.get(key));
+							put(key, ((Long)weapon_skills_json.get(key)).intValue());
 						}
 					}}
 					);
@@ -68,15 +68,15 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 			// Reads JSON file for all attributes of the heavy torso armor
 			// Overengineered, yes, but more memory-efficient
 			return new Torso_Armor("heavy", 
-					(Integer)torso_armor_json.get("BLUNT_RESISTANCE"),
-					(Integer)torso_armor_json.get("CUT_RESISTANCE"), 
-					(Integer)torso_armor_json.get("COVERAGE"), 
+					((Long)torso_armor_json.get("BLUNT_RESISTANCE")).intValue(),
+					((Long)torso_armor_json.get("CUT_RESISTANCE")).intValue(), 
+					((Long)torso_armor_json.get("COVERAGE")).intValue(), 
 					// Attributes
 					new HashMap<String, Integer> () {{
 						JSONObject attributes_json = (JSONObject)torso_armor_json.get("ATTRIBUTES");
 						for (Iterator it = attributes_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)attributes_json.get(key));
+							put(key, ((Long)attributes_json.get(key)).intValue());
 						}
 					}},
 					// Combat skills
@@ -84,7 +84,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject combat_skills_json = (JSONObject)torso_armor_json.get("COMBAT_SKILLS");
 						for (Iterator it = combat_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)combat_skills_json.get(key));
+							put(key, ((Long)combat_skills_json.get(key)).intValue());
 						}
 					}},
 					// Weapon skills
@@ -92,7 +92,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject weapon_skills_json = (JSONObject)torso_armor_json.get("WEAPON_SKILLS");
 						for (Iterator it = weapon_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)weapon_skills_json.get(key));
+							put(key, ((Long)weapon_skills_json.get(key)).intValue());
 						}
 					}}
 					);
@@ -119,15 +119,15 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 			// Reads JSON file for all attributes of the heavy leg armor
 			// Overengineered, yes, but more memory-efficient
 			return new Leg_Armor("heavy", 
-					(Integer)leg_armor_json.get("BLUNT_RESISTANCE"),
-					(Integer)leg_armor_json.get("CUT_RESISTANCE"), 
-					(Integer)leg_armor_json.get("COVERAGE"), 
+					((Long)leg_armor_json.get("BLUNT_RESISTANCE")).intValue(),
+					((Long)leg_armor_json.get("CUT_RESISTANCE")).intValue(), 
+					((Long)leg_armor_json.get("COVERAGE")).intValue(), 
 					// Attributes
 					new HashMap<String, Integer> () {{
 						JSONObject attributes_json = (JSONObject)leg_armor_json.get("ATTRIBUTES");
 						for (Iterator it = attributes_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)attributes_json.get(key));
+							put(key, ((Long)attributes_json.get(key)).intValue());
 						}
 					}},
 					// Combat skills
@@ -135,7 +135,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject combat_skills_json = (JSONObject)leg_armor_json.get("COMBAT_SKILLS");
 						for (Iterator it = combat_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)combat_skills_json.get(key));
+							put(key, ((Long)combat_skills_json.get(key)).intValue());
 						}
 					}},
 					// Weapon skills
@@ -143,7 +143,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject weapon_skills_json = (JSONObject)leg_armor_json.get("WEAPON_SKILLS");
 						for (Iterator it = weapon_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)weapon_skills_json.get(key));
+							put(key, ((Long)weapon_skills_json.get(key)).intValue());
 						}
 					}}
 					);
@@ -166,18 +166,18 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 		try {
 			final JSONObject boot_armor_json = (JSONObject)((JSONObject)parser.parse(new FileReader(this.stats_file))).get("BOOT_ARMOR");
 			
-			// Reads JSON file for all attributes of the heavy leg armor
+			// Reads JSON file for all attributes of the heavy boot armor
 			// Overengineered, yes, but more memory-efficient
 			return new Leg_Armor("heavy", 
-					(Integer)boot_armor_json.get("BLUNT_RESISTANCE"),
-					(Integer)boot_armor_json.get("CUT_RESISTANCE"), 
-					(Integer)boot_armor_json.get("COVERAGE"), 
+					((Long)boot_armor_json.get("BLUNT_RESISTANCE")).intValue(),
+					((Long)boot_armor_json.get("CUT_RESISTANCE")).intValue(), 
+					((Long)boot_armor_json.get("COVERAGE")).intValue(), 
 					// Attributes
 					new HashMap<String, Integer> () {{
 						JSONObject attributes_json = (JSONObject)boot_armor_json.get("ATTRIBUTES");
 						for (Iterator it = attributes_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)attributes_json.get(key));
+							put(key, ((Long)attributes_json.get(key)).intValue());
 						}
 					}},
 					// Combat skills
@@ -185,7 +185,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject combat_skills_json = (JSONObject)boot_armor_json.get("COMBAT_SKILLS");
 						for (Iterator it = combat_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)combat_skills_json.get(key));
+							put(key, ((Long)combat_skills_json.get(key)).intValue());
 						}
 					}},
 					// Weapon skills
@@ -193,7 +193,7 @@ public class Factory_Heavy_Armor implements Abstract_Factory_Armor {
 						JSONObject weapon_skills_json = (JSONObject)boot_armor_json.get("WEAPON_SKILLS");
 						for (Iterator it = weapon_skills_json.keySet().iterator(); it.hasNext();) {
 							String key = (String) it.next();
-							put(key, (Integer)weapon_skills_json.get(key));
+							put(key, ((Long)weapon_skills_json.get(key)).intValue());
 						}
 					}}
 					);
